Conflicts:
	radio/src/myeeprom.h
	radio/src/translations/cz.h
	radio/src/translations/de.h
	radio/src/translations/es.h
	radio/src/translations/fr.h
	radio/src/translations/fr.h.txt
	radio/src/translations/it.h
	radio/src/translations/pl.h
	radio/src/translations/pt.h
	radio/src/translations/se.h

+++ radio/src/gui/menus.h
@@ -291,7 +291,7 @@ int8_t switchMenuItem(uint8_t x, uint8_t y, int8_t value, LcdFlags attr, uint8_t
   #endif
   #define displayGVar(x, y, v, min, max) GVAR_MENU_ITEM(x, y, v, min, max, 0, 0, 0)
 #else
-  int8_t gvarMenuItem(uint8_t x, uint8_t y, int8_t value, int16_t min, int16_t max, LcdFlags attr, uint8_t event);
+    int16_t gvarMenuItem(uint8_t x, uint8_t y, int16_t value, int16_t min, int16_t max, LcdFlags attr, uint8_t event);
   #define displayGVar(x, y, v, min, max) lcd_outdez8(x, y, v)
 #endif
